+++ save-backend/src/main/kotlin/org/cqfn/save/backend/controllers/internal/UsersController.kt
@@ -2,7 +2,6 @@ package org.cqfn.save.backend.controllers.internal
 
 import org.cqfn.save.backend.repository.UserRepository
 import org.cqfn.save.backend.service.UserDetailsService
-import org.cqfn.save.backend.utils.IdentitySourceAwareUserDetails
 import org.cqfn.save.entities.User
 import org.slf4j.LoggerFactory
 import org.springframework.security.core.userdetails.UserDetails
@@ -13,7 +12,6 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import reactor.core.publisher.Mono
-import java.util.*
 
 /**
  * Controller that handles operation with users
