+++ save-cloud-common/src/jvmMain/kotlin/org/cqfn/save/utils/IdentitySourceAwareUserDetails.kt
@@ -1,0 +1,24 @@
+package org.cqfn.save.utils
+
+import org.springframework.security.core.authority.SimpleGrantedAuthority
+import org.springframework.security.core.userdetails.User
+
+/**
+ * @param authorities comma-separated authorities as a single string
+ * @property identitySource
+ * @property id
+ */
+class IdentitySourceAwareUserDetails(
+    username: String,
+    password: String?,
+    authorities: String?,
+    val identitySource: String,
+    val id: Long,
+) : User(
+    username,
+    password,
+    authorities?.split(',')
+        ?.filter { it.isNotBlank() }
+        ?.map { SimpleGrantedAuthority(it) }
+        ?: emptyList()
+)
